Relocate Prisma into packages/db (@troptix/db)
PR2 of the Prisma 7 plan (docs/plans/2026-06-prisma-7-upgrade.md, ADR 0012).
Moves the schema + client + config out of apps/web into the shared package, so
@troptix/db is the single source of truth for DB access. No schema/behavior
change.

packages/db (@troptix/db):
- Move schema.prisma + prisma.config.ts here; own the Prisma deps
  (@prisma/client, @prisma/adapter-pg, pg, prisma, @types/pg, dotenv);
  type:module; postinstall `prisma generate` → gitignored src/generated/.
- src/index.ts: server entry (server-only) — the pg-adapter singleton (moved
  from apps/web/src/server/prisma.ts) + re-exports the generated client surface
  (PrismaClient, Prisma namespace, enums, model types).
- src/types.ts: `export type *` from the generated client — RN-safe, runtime-free.

apps/web:
- src/server/prisma.ts → thin re-export of @troptix/db (call sites importing
  the singleton don't churn).
- Repoint 27 `@/generated/prisma/client` imports → `@troptix/db`.
- Drop the local Prisma deps + postinstall + schema/config; add @troptix/db as
  a workspace dependency.
- new-migration.ts points the CLI at packages/db's schema + `--config`; still
  runs from apps/web (where .env lives). apply-migration.ts unchanged.

Also folds in the deferred /simplify cleanup: shared package compilerOptions
hoisted into tsconfig.base.json (packages/{db,api}/tsconfig.json → extends +
include).

Validated: install generates @troptix/db's client; typecheck green (web +
@troptix/db); no-op `yarn db:new` → empty diff via the cross-package config;
`next build` succeeds; @troptix/db/types is runtime-free; server-only correctly
blocks a non-RSC import (the quarantine works).
