+++ apps/web/package.json
@@ -12,7 +12,6 @@
     "lint": "eslint .",
     "test": "jest",
     "test:watch": "jest --watch",
-    "postinstall": "prisma generate",
     "db:new": "tsx --env-file=.env scripts/new-migration.ts",
     "db:apply": "tsx --env-file=.env scripts/apply-migration.ts",
     "knip": "knip"
@@ -22,8 +21,6 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^5.0.1",
-    "@prisma/adapter-pg": "^7.8.0",
-    "@prisma/client": "^7.8.0",
     "@radix-ui/react-aspect-ratio": "^1.1.6",
     "@radix-ui/react-avatar": "^1.1.6",
     "@radix-ui/react-checkbox": "^1.2.3",
@@ -49,6 +46,7 @@
     "@stripe/stripe-js": "^2.2.0",
     "@tanstack/react-query": "^5.8.4",
     "@tanstack/react-table": "^8.21.3",
+    "@troptix/db": "0.0.0",
     "@vercel/analytics": "^1.1.1",
     "@vercel/node": "^5.1.14",
     "@vercel/toolbar": "^0.1.34",
@@ -71,7 +69,6 @@
     "micro": "^10.0.1",
     "motion": "^12.23.25",
     "next": "16.0.10",
-    "pg": "^8.13.0",
     "postcss": "8.4.29",
     "posthog-js": "^1.255.1",
     "posthog-node": "^5.1.1",
@@ -98,7 +95,6 @@
     "@testing-library/react": "^14.2.0",
     "@types/jest": "^29.5.12",
     "@types/node": "^24.10.1",
-    "@types/pg": "^8.11.10",
     "@types/qrcode.react": "^3.0.0",
     "@types/react": "19.2.7",
     "@types/react-dom": "19.2.3",
@@ -107,7 +103,6 @@
     "eslint-config-next": "16.0.7",
     "knip": "^5.62.0",
     "ngrok": "^5.0.0-beta.2",
-    "prisma": "^7.8.0",
     "qrcode": "^1.5.4",
     "react-email": "4.1.3",
     "tailwindcss": "3.3.3",
