+++ apps/web/src/app/api/checkout/config/route.ts
@@ -2,11 +2,7 @@
 
 import { NextRequest, NextResponse } from 'next/server';
 import prisma from '@/server/prisma';
-import {
-  OrderStatus,
-  Prisma,
-  TicketFeeStructure,
-} from '@/generated/prisma/client';
+import { OrderStatus, Prisma, TicketFeeStructure } from '@troptix/db';
 import { calculateFees } from '@/lib/fees';
 import { CheckoutConfigResponse, CheckoutTicket } from '@/types/checkout';
