+++ apps/web/src/app/api/checkout/initiate/route.ts
@@ -4,11 +4,7 @@ import { cookies, headers } from 'next/headers'; // For reading cookies in App R
 import prisma from '@/server/prisma';
 import { stripe } from '@/server/lib/stripe';
 import { UserDetailsFormData } from '@/lib/schemas/checkoutSchema'; // Adjust path if needed
-import {
-  OrderStatus,
-  TicketFeeStructure,
-  TicketStatus,
-} from '@/generated/prisma/client';
+import { OrderStatus, TicketFeeStructure, TicketStatus } from '@troptix/db';
 import { getFirebaseAdmin } from '@/server/lib/firebaseAdmin'; // Adjust path if needed
 import { generateId } from '@/lib/utils'; // Adjust path if needed
 import {
@@ -18,7 +14,7 @@ import {
   ValidationResponseMessage,
 } from '@/types/checkout'; // Adjust path if needed
 import { calculateFees } from '@/lib/fees'; // Adjust path if needed
-import { Prisma } from '@/generated/prisma/client';
+import { Prisma } from '@troptix/db';
 import { sendEmailConfirmationEmailToUser } from '@/server/lib/email';
 
 interface InitiateRequestData {
