+++ apps/web/src/app/api/organizer/tickets/scan/route.ts
@@ -1,6 +1,6 @@
 import { getUserFromIdTokenCookie } from '@/server/authUser';
 import prisma from '@/server/prisma';
-import { TicketStatus } from '@/generated/prisma/client';
+import { TicketStatus } from '@troptix/db';
 import { headers } from 'next/headers';
 import { NextRequest, NextResponse } from 'next/server';
