+++ apps/web/src/app/organizer/_lib/getEventsData.ts
@@ -1,4 +1,4 @@
-import { Prisma } from '@/generated/prisma/client';
+import { Prisma } from '@troptix/db';
 
 import prisma from '@/server/prisma';
