+++ apps/web/src/app/organizer/events/[eventId]/attendees/_actions/attendeeActions.ts
@@ -3,7 +3,7 @@
 import { revalidatePath } from 'next/cache';
 import prisma from '@/server/prisma';
 import { getUserFromIdTokenCookie } from '@/server/authUser';
-import { TicketStatus } from '@/generated/prisma/client';
+import { TicketStatus } from '@troptix/db';
 import { getEventWhereClause, verifyEventAccess } from '@/server/accessControl';
 
 export async function toggleTicketStatus(ticketId: string, eventId: string) {
