+++ apps/web/src/server/lib/promotionHelper.ts
@@ -1,4 +1,4 @@
-import { Prisma } from '@/generated/prisma/client';
+import { Prisma } from '@troptix/db';
 
 export function getPrismaUpdatePromotionQuery(promo) {
   let promotion: Prisma.PromotionsUpdateInput;
