+++ apps/web/src/server/lib/ticketHelper.ts
@@ -1,4 +1,4 @@
-import { Prisma, PrismaClient, TicketStatus } from '@/generated/prisma/client';
+import { Prisma, PrismaClient, TicketStatus } from '@troptix/db';
 import prisma from '@/server/prisma';
 
 const prismaClient = prisma as PrismaClient;
