+++ packages/db/package.json
@@ -2,7 +2,8 @@
   "name": "@troptix/db",
   "version": "0.0.0",
   "private": true,
-  "description": "TropTix shared database layer (Drizzle schema + client). Server-only; RN-safe inferred types under ./types.",
+  "type": "module",
+  "description": "TropTix shared database layer (Prisma 7 schema + client). Server-only; RN-safe types under ./types.",
   "main": "./src/index.ts",
   "types": "./src/index.ts",
   "exports": {
@@ -16,12 +17,20 @@
     }
   },
   "scripts": {
-    "typecheck": "tsc --noEmit"
+    "typecheck": "tsc --noEmit",
+    "postinstall": "prisma generate",
+    "generate": "prisma generate"
   },
   "dependencies": {
+    "@prisma/adapter-pg": "^7.8.0",
+    "@prisma/client": "^7.8.0",
+    "pg": "^8.13.0",
     "server-only": "^0.0.1"
   },
   "devDependencies": {
+    "@types/pg": "^8.11.10",
+    "dotenv": "^17.2.0",
+    "prisma": "^7.8.0",
     "typescript": "^5.2.2"
   }
 }
