+++ packages/db/src/index.ts
@@ -1,13 +1,47 @@
 // @troptix/db — SERVER ENTRY.
 //
-// This entry owns server-only runtime: the Drizzle client, `pg`, and anything
-// that must never reach a client/RN bundle. The `server-only` import makes a
-// build fail loudly if this module is ever pulled into a client bundle.
-//
-// Stage 0 is wiring only — the Drizzle schema/client land in Stage 1
-// (see docs/plans/2026-06-shared-packages-platform.md). For now this is an
-// intentional placeholder so the workspace graph and the server/types split
-// resolve.
+// Owns the Prisma 7 runtime: the generated client + the `pg` driver adapter.
+// `import 'server-only'` makes a build fail loudly if this is ever pulled into
+// a client/RN bundle. Clients import the type-only `@troptix/db/types` entry.
 import 'server-only';
+import { PrismaPg } from '@prisma/adapter-pg';
+import { PrismaClient } from './generated/prisma/client';
 
-export {};
+/**
+ * Prisma 7 runtime client (moved here from apps/web in the packages/db
+ * relocation). The Rust engine is gone — the client connects through the `pg`
+ * driver adapter, with the connection URL read from the environment of the
+ * consuming app (Next provides `POSTGRES_PRISMA_URL`, pooled). See
+ * docs/plans/2026-06-prisma-7-upgrade.md for the pgbouncer / SSL notes.
+ *
+ * A single client is cached on `globalThis` in dev so HMR doesn't open a new
+ * pool per reload.
+ */
+const createPrismaClient = () =>
+  new PrismaClient({
+    adapter: new PrismaPg({
+      connectionString: process.env.POSTGRES_PRISMA_URL,
+      // node-pg has no default connection timeout; Prisma v6 used 5s.
+      connectionTimeoutMillis: 5000,
+      // Match pre-v7 SSL behavior (the Rust engine ignored cert validation).
+      ssl: { rejectUnauthorized: false },
+    }),
+  });
+
+const globalForPrisma = globalThis as unknown as {
+  prisma: PrismaClient | undefined;
+};
+
+const prisma = globalForPrisma.prisma ?? createPrismaClient();
+
+if (process.env.NODE_ENV !== 'production') {
+  globalForPrisma.prisma = prisma;
+}
+
+export default prisma;
+export { prisma };
+
+// Re-export the generated client's surface (PrismaClient, the `Prisma`
+// namespace, enums, model types) so server consumers import everything from
+// `@troptix/db` rather than reaching into the generated output directly.
+export * from './generated/prisma/client';
