+++ tsconfig.base.json
@@ -1,7 +1,20 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
+  // Shared baseline for workspace packages (packages/*). The apps
+  // (apps/web, apps/organizer) intentionally do NOT extend this — they carry
+  // their own framework-specific compiler settings — but they reuse the same
+  // @troptix/* path aliases declared here.
   "compilerOptions": {
     "baseUrl": ".",
+    "target": "ES2022",
+    "module": "ESNext",
+    "moduleResolution": "Bundler",
+    "lib": ["ES2022"],
+    "strict": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "noEmit": true,
+    "types": ["node"],
     "paths": {
       "@troptix/db": ["packages/db/src/index.ts"],
       "@troptix/db/types": ["packages/db/src/types.ts"],
